Refactoring: Add NVS manager for persistent storage and button state handling
