+++ include/commandHandler.h
@@ -19,5 +19,3 @@ void handleCommand(uint8_t commandType, uint8_t value);
 void setAmpChannel(uint8_t channel);
 void checkAmpChannelButtons();
 void handleProgramChange(byte midiChannel, byte program);
-void saveMidiChannelToNVS();
-void loadMidiChannelFromNVS();
