+++ platformio.ini
@@ -17,11 +17,11 @@ build_flags =
 	-D ARDUINO_USB_CDC_ON_BOOT=1
 	-D ARDUINO_USB_MODE=1
 	-D CLIENT_TYPE=AMP_SWITCHER
-	-D MAX_AMPSWITCHS=1
-	;-D AMP_SWITCH_PINS=\"2,9,10,20\"
-	;-D AMP_BUTTON_PINS=\"1,3,4,5\"   ; Channel 1=8, 2=9, 3=10, 4=20
-	-D AMP_SWITCH_PINS=\"2\"
-	-D AMP_BUTTON_PINS=\"1\"   ; Channel 1=8, 2=9, 3=10, 4=20
+	-D MAX_AMPSWITCHS=4
+	-D AMP_SWITCH_PINS=\"2,9,10,20\"
+	-D AMP_BUTTON_PINS=\"1,3,4,5\"   ; Channel 1=8, 2=9, 3=10, 4=20
+	;-D AMP_SWITCH_PINS=\"2\"
+	;-D AMP_BUTTON_PINS=\"1\"   ; Channel 1=8, 2=9, 3=10, 4=20
 	-D DEVICE_NAME=\"AMP_SWITCHER_1\"
 lib_compat_mode = soft
 lib_ldf_mode = chain
